🤶🏼 User token (#702)
* fix error of token when user not login

* fix error of token when user not login

* fix token error for build profile

* fix token error for build profile

* fix token error for build profile

* fix email template of app_approved and build_profile

When users APPLY, they receive an email once APPROVED by Admin, I need that email to have a direct link to the user trainings, and after they complete the training receive another email, I need that email too with a direct link so that they build the profile.

* add log file app.log

* add log file app.log

* restore app.log

* #682

2. Do not allow duplicated values using same email in collections in database, if anyone is applying with same email for an existing role in the platform, we should not allow it, should be informed in a text message to the user saying that User already exists.

* #701

reduce the size of the initial boxes and people login

You see it as this below when opening it. Could we reduce the space inbetween the text lines and then make the boxes a bit smaller to all fit on the screen.